# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "server",
    "description": "",
    "license": {
      "name": "AGPL-3.0-only",
      "identifier": "AGPL-3.0-only"
    },
    "version": "0.0.1"
  },
  "paths": {
...
    "/api/performances/{id}": {
      "get": {
        "tags": [
          "performances"
        ],
        "operationId": "get_performance",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Performance ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Performance detail",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PerformanceResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "performances"
        ],
        "operationId": "update_performance",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Performance ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePerformanceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Updated performance",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PerformanceResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      },
      "delete": {
        "tags": [
          "performances"
        ],
        "operationId": "delete_performance",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Performance ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
    "/api/performances/{id}/audio": {
      "post": {
        "tags": [
          "performances"
        ],
        "operationId": "upload_audio",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Performance ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/AudioUpload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Audio uploaded and linked",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AudioInfo"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Performance not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
    "/api/performances/{id}/audio/link": {
      "post": {
        "tags": [
          "performances"
        ],
        "operationId": "add_audio_link",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Performance ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddAudioLinkRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Audio link created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AudioInfo"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Performance not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
    "/api/performances/{id}/audio/{asset_id}": {
      "delete": {
        "tags": [
          "performances"
        ],
        "operationId": "delete_audio",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Performance ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "asset_id",
            "in": "path",
            "description": "Audio asset ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Unlinked"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      },
      "patch": {
        "tags": [
          "performances"
        ],
        "operationId": "update_audio_kind",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Performance ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "asset_id",
            "in": "path",
            "description": "Audio asset ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAudioKindRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Kind updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AudioInfo"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "AddAudioLinkRequest": {
        "type": "object",
        "description": "Request body for `POST /api/performances/{id}/audio/link`.",
        "required": [
          "external_url",
          "kind"
        ],
        "properties": {
          "credits": {
            "type": [
              "string",
              "null"
            ],
            "description": "Attribution text for the content creator."
          },
          "external_url": {
            "type": "string",
            "description": "The external audio URL to reference."
          },
          "kind": {
            "type": "string",
            "description": "Serialized [`AudioKind`] value to assign."
          },
          "source_url": {
            "type": [
              "string",
              "null"
            ],
            "description": "Link to the original source post or stream."
          },
          "title": {
            "type": [
              "string",
              "null"
            ],
            "description": "Display label for this link."
          }
        }
      },
...
      "ArtistInfo": {
        "type": "object",
        "description": "A minimal artist record embedded in song and performance responses.",
        "required": [
          "id",
          "name"
        ],
        "properties": {
          "description": {
            "type": [
              "string",
              "null"
            ],
            "description": "Optional freeform bio or description."
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Unique identifier."
          },
          "name": {
            "type": "string",
            "description": "Display name."
          }
        }
      },
...
      "AudioInfo": {
        "allOf": [
          {
            "$ref": "#/components/schemas/AssetInfo"
          },
          {
            "type": "object",
            "required": [
              "kind"
            ],
            "properties": {
              "kind": {
                "type": "string",
                "description": "Serialized [`AudioKind`] value."
              }
            }
          }
        ],
        "description": "An audio asset attached to a performance with its semantic role."
      },
...
      "AudioUpload": {
        "type": "object",
        "description": "Placeholder schema for audio multipart upload bodies.",
        "required": [
          "file",
          "kind"
        ],
        "properties": {
          "credits": {
            "type": [
              "string",
              "null"
            ]
          },
          "file": {
            "type": "string",
            "format": "binary"
          },
          "kind": {
            "type": "string",
            "description": "Semantic role. See [`AudioKind`]."
          },
          "source_url": {
            "type": [
              "string",
              "null"
            ]
          },
          "title": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "description": "Body returned for all error responses.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "Human readable description of the error."
          }
        }
      },
...
      "PerformanceResponse": {
        "type": "object",
        "description": "Full performance metadata returned by detail endpoints.\n\nExcludes lyrics, fetch those via `GET /api/performances/{id}/lyrics` on demand.",
        "required": [
          "id",
          "play_count",
          "performance_date",
          "stream_number",
          "performance_number",
          "songs",
          "singers",
          "tags",
          "audio",
          "video"
        ],
        "properties": {
          "audio": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AudioInfo"
            },
            "description": "Audio files attached to this performance."
          },
          "duration": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "description": "Duration in seconds.",
            "minimum": 0
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Unique identifier."
          },
          "performance_date": {
            "type": "string",
            "format": "date",
            "description": "Calendar date of the stream."
          },
          "performance_number": {
            "type": "integer",
            "format": "int32",
            "description": "Performance index within `performance_date` + `stream_number`, 1-based.",
            "minimum": 0
          },
          "play_count": {
            "type": "integer",
            "format": "int32",
            "description": "Number of times this performance has been played."
          },
          "singers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ArtistInfo"
            },
            "description": "Artists who sang in this performance."
          },
          "songs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SongSummary"
            },
            "description": "Songs featured in this performance."
          },
          "stream_number": {
            "type": "integer",
            "format": "int32",
            "description": "Stream index within `performance_date`, 1-based.",
            "minimum": 0
          },
          "stream_time": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "description": "Offset in seconds from the start of the stream.",
            "minimum": 0
          },
          "tags": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TagInfo"
            },
            "description": "Tags applied to this performance."
          },
          "title": {
            "type": [
              "string",
              "null"
            ],
            "description": "Optional display title. Absent when no explicit title has been set."
          },
          "video": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VideoInfo"
            },
            "description": "Video files attached to this performance."
          }
        }
      },
...
      "TagInfo": {
        "type": "object",
        "description": "A tag embedded in song responses.",
        "required": [
          "id",
          "name",
          "kind"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Unique identifier."
          },
          "kind": {
            "type": "string",
            "description": "Freeform category string (e.g. `\"genre\"`, `\"mood\"`)."
          },
          "name": {
            "type": "string",
            "description": "Display name of the tag."
          }
        }
      },
...
      "UpdateAudioKindRequest": {
        "type": "object",
        "description": "Request body for `PATCH /api/performances/{id}/audio/{asset_id}`.",
        "required": [
          "kind"
        ],
        "properties": {
          "kind": {
            "type": "string",
            "description": "Serialized [`AudioKind`] value to assign."
          }
        }
      },
...
      "UpdatePerformanceRequest": {
        "type": "object",
        "description": "Request body for `PUT /api/performances/{id}`.\n\nLyrics are excluded, use `PUT /api/performances/{id}/lyrics` instead.",
        "required": [
          "performance_date",
          "stream_number",
          "performance_number",
          "song_ids",
          "singer_ids",
          "tags"
        ],
        "properties": {
          "duration": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "description": "Duration in seconds.",
            "minimum": 0
          },
          "performance_date": {
            "type": "string",
            "format": "date",
            "description": "Calendar date of the stream in which this performance occurred."
          },
          "performance_number": {
            "type": "integer",
            "format": "int32",
            "description": "Performance index within `performance_date` + `stream_number`, 1-based.",
            "minimum": 0
          },
          "singer_ids": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "description": "IDs of artists who sang in this performance."
          },
          "song_ids": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "description": "IDs of songs featured in this performance."
          },
          "stream_number": {
            "type": "integer",
            "format": "int32",
            "description": "Stream index within `performance_date`, 1-based.",
            "minimum": 0
          },
          "stream_time": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "description": "Offset in seconds from the start of the stream.",
            "minimum": 0
          },
          "tags": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PerformanceTagAssignment"
            },
            "description": "Replaces all existing tags."
          },
          "title": {
            "type": [
              "string",
              "null"
            ],
            "description": "Optional display title override."
          }
        }
      },
...
      "VideoInfo": {
        "allOf": [
          {
            "$ref": "#/components/schemas/AssetInfo"
          },
          {
            "type": "object",
            "required": [
              "kind"
            ],
            "properties": {
              "kind": {
                "type": "string",
                "description": "Serialized [`VideoKind`] value."
              }
            }
          }
        ],
        "description": "A video asset attached to a performance with its semantic role."
      },
...
    },
    "securitySchemes": {
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session"
      }
    }
  }
}